+++ src/containers/WebComponentLoader.test.js
@@ -29,7 +29,11 @@ let cookies;
 const code = "print('This project is amazing')";
 const identifier = "My amazing project";
 const steps = [{ quiz: false, title: "Step 1", content: "Do something" }];
-const instructions = { currentStepPosition: 3, project: { steps: steps } };
+const instructions = {
+  currentStepPosition: 3,
+  project: { steps: steps },
+  permitOverride: false,
+};
 const authKey = "my_key";
 const user = { access_token: "my_token" };